10/23-2

+++ wav_program/Signal.h
@@ -6,7 +6,7 @@
 class Signal
 {
 public:
-	unsigned int Fs;
+	double Fs;
 	std::vector<double> dataL;
 	std::vector<double> dataR;
